### 2020-02-11 #2093 @whedon
[PRE REVIEW]: PySINDy: A Python package for the sparse identification of nonlinear dynamical systems from data
**Submitting author:** @briandesilva (<a href="http://orcid.org/0000-0003-0944-900X">Brian de Silva</a>)
**Repository:** <a href="https://github.com/dynamicslab/pysindy" target ="_blank">https://github.com/dynamicslab/pysindy</a>
**Version:** v0.12.0
