### 2020-02-11 #2093 @danielskatz
@briandesilva - it looks like you may need to add a bunch of DOIs to your references (although whedon isn't always 100% correct on its guesses) - see the [example submission and references](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) for how to add DOIs.  Once you have done this, regenerate the paper by entering ```@whedon generate pdf``` as a new comment here
