### 2020-02-11 cac41ad
Set nonzero default regularization for STLSQ.
