### 2020-02-11 6d6de6c
Add DOIs to references.
