+++ docs/JOSS/paper.bib
@@ -1,29 +1,31 @@
 @article{brunton2016pnas,
-  title={Discovering governing equations from data by sparse identification of nonlinear dynamical systems},
-  author={Brunton, Steven L. and Proctor, Joshua L. and Kutz, J. Nathan},
-  journal={Proceedings of the National Academy of Sciences},
-  volume={113},
-  number={15},
-  pages={3932--3937},
-  year={2016},
-  publisher={National Academy of Sciences}
+  title = {Discovering governing equations from data by sparse identification of nonlinear dynamical systems},
+  author = {Brunton, Steven L. and Proctor, Joshua L. and Kutz, J. Nathan},
+  journal = {Proceedings of the National Academy of Sciences},
+  volume = {113},
+  number = {15},
+  pages = {3932--3937},
+  year = {2016},
+  publisher = {National Academy of Sciences},
+  doi = {10.1073/pnas.1517384113}
 }
 
 @article{zheng2018ieee,
-  title={A unified framework for sparse relaxed regularized regression: SR3},
-  author={Zheng, Peng and Askham, Travis and Brunton, Steven L. and Kutz, J. Nathan and Aravkin, Aleksandr Y.},
-  journal={IEEE Access},
-  volume={7},
-  pages={1404--1423},
-  year={2018},
-  publisher={IEEE}
+  title = {A unified framework for sparse relaxed regularized regression: SR3},
+  author = {Zheng, Peng and Askham, Travis and Brunton, Steven L. and Kutz, J. Nathan and Aravkin, Aleksandr Y.},
+  journal = {IEEE Access},
+  volume = {7},
+  pages = {1404--1423},
+  year = {2018},
+  publisher = {IEEE},
+  doi = {10.1109/access.2018.2886528}
 }
 
 @article{champion2019,
-  title={A unified sparse optimization framework to learn parsimonious physics-informed models from data},
-  author={Champion, Kathleen and Zheng, Peng and Aravkin, Aleksandr Y. and Brunton, Steven L. and Kutz, J. Nathan},
-  journal={arXiv preprint arXiv:1906.10612},
-  year={2019}
+  title = {A unified sparse optimization framework to learn parsimonious physics-informed models from data},
+  author = {Champion, Kathleen and Zheng, Peng and Aravkin, Aleksandr Y. and Brunton, Steven L. and Kutz, J. Nathan},
+  journal = {arXiv preprint arXiv:1906.10612},
+  year = {2019}
 }
 
 @article {Schmidt81,
@@ -35,7 +37,8 @@
   year = {2009},
   publisher = {American Association for the Advancement of Science},
   issn = {0036-8075},
-  journal = {Science}
+  journal = {Science},
+  doi = {10.1126/science.1165893}
 }
 
 @article{PhysRevMaterials.2.083802,
@@ -47,7 +50,8 @@
   pages = {083802},
   numpages = {11},
   year = {2018},
-  publisher = {American Physical Society}
+  publisher = {American Physical Society},
+  doi = {10.1103/physrevmaterials.2.083802}
 }
 
 @article{10.2307/2346178,
@@ -59,7 +63,8 @@
  publisher = {Royal Statistical Society, Wiley},
  title = {Regression Shrinkage and Selection via the Lasso},
  volume = {58},
- year = {1996}
+ year = {1996},
+ doi = {10.1111/j.2517-6161.1996.tb02080.x}
 }
 
 @misc{markus_quade_sparsereg,
@@ -67,6 +72,7 @@
[…]
