### 2020-02-12 41863da
Removed unbiasing from BaseOptimizer
