### 2020-02-12 de89ff2
Added new SINDyOptimizer object as a wrapper that performs unbiasing. Modified the SINDy class to make use of this new class in the pipeline
