### 2020-02-12 f2d34e1
Fixed issue in equations function. Checks if the intercept is a scalar (happens when fit_intercept is False) or a vectorand use this accordingly in forming the model equations
