+++ pysindy/utils/base.py
@@ -142,8 +142,10 @@ def equations(pipeline, input_features=None, precision=3, input_fmt=None):
         input_features = [input_fmt(i) for i in input_features]
     coef = pipeline.steps[-1][1].coef_
     intercept = pipeline.steps[-1][1].intercept_
+    if np.isscalar(intercept):
+        intercept = intercept*np.ones(coef.shape[0])
     return [
-        print_model(coef[i], input_features, intercept=intercept, precision=precision)
+        print_model(coef[i], input_features, intercept=intercept[i], precision=precision)
         for i in range(coef.shape[0])
     ]
 
