### 2020-02-12 e9b393a
added SINDyOptimizer to list of imports in __init__.py
