+++ pysindy/optimizers/__init__.py
@@ -1,4 +1,5 @@
 from .base import BaseOptimizer
+from .sindy_optimizer import SINDyOptimizer
 from .elastic_net import ElasticNet
 from .lasso import LASSO
 from .sr3 import SR3
