+++ pysindy/optimizers/__init__.py
@@ -1,6 +1,6 @@
 from .base import BaseOptimizer
-from .sindy_optimizer import SINDyOptimizer
 from .elastic_net import ElasticNet
 from .lasso import LASSO
+from .sindy_optimizer import SINDyOptimizer
 from .sr3 import SR3
 from .stlsq import STLSQ
