### 2020-02-12 a971790
renaming parameter passed to SINDyOptimizer
