### 2020-02-12 1221381
Updated complexity computation in the SINDyOptimizer class
