### 2020-02-12 a10d655
updated SINDyOptimizer to comply with BaseEstimator properties (enabling check_is_fitted to work properly)
