### 2020-02-12 386d52b
Updating unit tests to use sklearn version of Lasso and ElasticNet
