+++ test/optimizers/test_optimizers.py
@@ -3,23 +3,27 @@ Unit tests for optimizers.
 """
 import pytest
 from numpy.linalg import norm
+from sklearn.linear_model import ElasticNet
+from sklearn.linear_model import Lasso
 from sklearn.utils.validation import check_is_fitted
 
-from pysindy.optimizers import ElasticNet
-from pysindy.optimizers import LASSO
 from pysindy.optimizers import SINDyOptimizer
 from pysindy.optimizers import SR3
 from pysindy.optimizers import STLSQ
 
 
-@pytest.mark.parametrize("optimizer", [STLSQ(), SR3(), LASSO(), ElasticNet()])
+@pytest.mark.parametrize(
+    "optimizer",
+    [STLSQ(), SR3(), Lasso(fit_intercept=False), ElasticNet(fit_intercept=False)],
+)
 def test_fit(data_derivative_1d, optimizer):
     x, x_dot = data_derivative_1d
     x = x.reshape(-1, 1)
-    optimizer.fit(x, x_dot)
+    opt = SINDyOptimizer(optimizer, unbias=False)
+    opt.fit(x, x_dot)
 
-    check_is_fitted(optimizer)
-    assert optimizer.complexity >= 0
+    check_is_fitted(opt)
+    assert opt.complexity >= 0
 
 
 @pytest.mark.parametrize(
@@ -63,21 +67,6 @@ def test_bad_parameters(data_derivative_1d):
     with pytest.raises(ValueError):
         SR3(max_iter=0)
 
-    with pytest.raises(ValueError):
-        LASSO(alpha=-1)
-
-    with pytest.raises(ValueError):
-        LASSO(max_iter=0)
-
-    with pytest.raises(ValueError):
-        ElasticNet(alpha=-1)
-
-    with pytest.raises(ValueError):
-        ElasticNet(max_iter=0)
-
-    with pytest.raises(ValueError):
-        ElasticNet(l1_ratio=-0.1)
-
 
 # The different capitalizations are intentional;
 # I want to make sure different versions are recognized
@@ -90,18 +79,21 @@ def test_sr3_prox_functions(data_derivative_1d, thresholder):
     check_is_fitted(model)
 
 
+@pytest.mark.parametrize(
+    "optimizer",
+    [
+        STLSQ(threshold=0.01, alpha=0.1, max_iter=1),
+        Lasso(alpha=0.1, fit_intercept=False, max_iter=1),
+    ],
+)
 def test_unbias(data_derivative_1d):
     x, x_dot = data_derivative_1d
     x = x.reshape(-1, 1)
 
-    optimizer_biased = SINDyOptimizer(
-        STLSQ(threshold=0.01, alpha=0.01, max_iter=1), unbias=False
-    )
+    optimizer_biased = SINDyOptimizer(optimizer, unbias=False)
     optimizer_biased.fit(x, x_dot)
 
-    optimizer_unbiased = SINDyOptimizer(
-        STLSQ(threshold=0.01, alpha=0.01, max_iter=1), unbias=True
-    )
+    optimizer_unbiased = SINDyOptimizer(optimizer, unbias=True)
     optimizer_unbiased.fit(x, x_dot)
 
     assert (
