+++ test/optimizers/test_optimizers.py
@@ -79,21 +79,31 @@ def test_sr3_prox_functions(data_derivative_1d, thresholder):
     check_is_fitted(model)
 
 
-@pytest.mark.parametrize(
-    "optimizer",
-    [
-        STLSQ(threshold=0.01, alpha=0.1, max_iter=1),
-        Lasso(alpha=0.1, fit_intercept=False, max_iter=1),
-    ],
-)
-def test_unbias(data_derivative_1d, optimizer):
+def test_unbias(data_derivative_1d):
     x, x_dot = data_derivative_1d
     x = x.reshape(-1, 1)
 
-    optimizer_biased = SINDyOptimizer(optimizer, unbias=False)
+    optimizer_biased = SINDyOptimizer(STLSQ(threshold=0.01, alpha=0.1, max_iter=1), unbias=False)
     optimizer_biased.fit(x, x_dot)
 
-    optimizer_unbiased = SINDyOptimizer(optimizer, unbias=True)
+    optimizer_unbiased = SINDyOptimizer(STLSQ(threshold=0.01, alpha=0.1, max_iter=1), unbias=True)
+    optimizer_unbiased.fit(x, x_dot)
+
+    assert (
+        norm(optimizer_biased.coef_ - optimizer_unbiased.coef_)
+        / norm(optimizer_unbiased.coef_)
+        > 1e-9
+    )
+
+
+def test_unbias_external(data_derivative_1d):
+    x, x_dot = data_derivative_1d
+    x = x.reshape(-1, 1)
+
+    optimizer_biased = SINDyOptimizer(Lasso(alpha=0.1, fit_intercept=False, max_iter=1), unbias=False)
+    optimizer_biased.fit(x, x_dot)
+
+    optimizer_unbiased = SINDyOptimizer(Lasso(alpha=0.1, fit_intercept=False, max_iter=1), unbias=True)
     optimizer_unbiased.fit(x, x_dot)
 
     assert (
