### 2020-02-12 1245fdf
fixing case for when external optimizer returns prediction as 1D array instead of 2D array
