### 2020-02-12 2c57f03
removing our implementations of ElasticNet and Lasso
