### 2020-02-12 8c6d012
removing LASSO and ElasticNet from __init__.py file
