+++ pysindy/optimizers/__init__.py
@@ -1,6 +1,4 @@
 from .base import BaseOptimizer
-from .elastic_net import ElasticNet
-from .lasso import LASSO
 from .sindy_optimizer import SINDyOptimizer
 from .sr3 import SR3
 from .stlsq import STLSQ
