### 2020-02-12 850e409
add test for supports_multiple_targets
