### 2020-02-12 4267372
move multitarget check to SINDyOptimizer abstraction layer
