+++ pysindy/pysindy.py
@@ -16,10 +16,8 @@ from sklearn.utils.validation import check_is_fitted
 from pysindy.differentiation import FiniteDifference
 from pysindy.optimizers import SINDyOptimizer
 from pysindy.optimizers import STLSQ
-from pysindy.optimizers.base import _MultiTargetLinearRegressor
 from pysindy.utils.base import drop_nan_rows
 from pysindy.utils.base import equations
-from pysindy.utils.base import supports_multiple_targets
 from pysindy.utils.base import validate_input
 
 
@@ -168,11 +166,7 @@ class SINDy(BaseEstimator):
         # Drop rows where derivative isn't known
         x, x_dot = drop_nan_rows(x, x_dot)
 
-        if supports_multiple_targets(self.optimizer):
-            optimizer = self.optimizer
-        else:
-            optimizer = _MultiTargetLinearRegressor(self.optimizer)
-        optimizer = SINDyOptimizer(optimizer, unbias=unbias)
+        optimizer = SINDyOptimizer(self.optimizer, unbias=unbias)
         steps = [("features", self.feature_library), ("model", optimizer)]
         self.model = Pipeline(steps)
 
