### 2020-02-14 a33da62
Add JOSS badge to README
