### 2020-02-14 3a2fbdd
Update unbias in optimizer docstrings.
