+++ pysindy/optimizers/base.py
@@ -49,15 +49,6 @@ class BaseOptimizer(LinearRegression, ComplexityMixin):
 
     copy_X : boolean, optional (default True)
         If True, X will be copied; else, it may be overwritten.
-
-    unbias : boolean, optional (default True)
-        Whether to perform an extra step of unregularized linear regression to unbias
-        the coefficients for the identified support.
-        For example, if `STLSQ(alpha=0.1)` is used then the learned coefficients will
-        be biased toward 0 due to the L2 regularization.
-        Setting `unbias=True` will trigger an additional step wherein the nonzero
-        coefficients learned by the `STLSQ` object will be updated using an
-        unregularized least-squares fit.
     """
 
     def __init__(self, max_iter=20, normalize=False, fit_intercept=False, copy_X=True):
