### 2020-02-19 9069ea3
Update sklearn requirements.
We should have been requiring sklearn 0.21 since the `PolynomialLibrary` object imports a function from sklearn not present in earlier versions. See #51 .
