### 2020-02-20 db8ef41
added checks for the various methods and attributes that the optimizer passed into a SINDyOptimizer should use
