### 2020-02-20 b5fa7f9
fixed bug in checking optimizer for coefficient attribute
