### 2020-03-06 1a38592
Add documentation to PolynomialLibrary.
