### 2020-03-07 #2104 @sixpearls
Hi @briandesilva et al, this is a very nice library! I checked most of the items in the list, but I did have a few comments, particularly on the unchecked items, below.

## Authorship
I am not sure I see the SOFTWARE contribution of the last three authors, although it seems that based on the references these authors may have contributed to the theoretical development. In the past, I believe the JOSS policy was relatively strict about the software contribution of each author. Is that still the case? Could you please advise @terrytangyuan?

# Documentation:

## Statement of need

The documentation itself does not provide any statement of need or description of the mathematical formulation. I understand that the references that describe the problem are open-access, however in my experience there was a significant cognitive burden to understand the purpose of the library. These are my recommendations:
- [ ] At the very least, please provide a prose reference to the papers in the readme (as currently done in the top of the `sindy_paper.ipynb` example notebook).
- [ ] It would be nice to embed the example notebooks in the docs -- this would reduce the cognitive burden for understanding the purpose of the library prior to installation and (I believe sphinx does this automatically:) make it easier to download the notebook so users can play with them.
- [ ] It would be nice if a brief overview of the mathematics was included in at least one of the example notebooks; one particular benefit of using notebooks for examples is that it is easy to interweave the mathematical concepts with the code that implements it and that is essentially absent. For example, walking through some of the [supplemental examples](https://www.pnas.org/content/pnas/suppl/2016/03/23/1517384113.DCSupplemental/pnas.1517384113.sapp.pdf) or really the majority of the [unified framework paper](https://arxiv.org/pdf/1906.10612.pdf) including the equations and software implementation would be really nice. This could especially improve the API documentation for tuning the basis function family, optimizer, and other parameters, which is currently a bit difficult to parse through.
- [ ] It would greatly improve the experience for the reader to include at least some mathematical discussion and example code in the readme and setup.py `long_description` so that the benefits of the previous two suggests can be obtained for users without leaving github or PyPI.

## Tests

I have checked it off because there does seem to be a thorough set of tests that cover the API, but it would improve the library to include regression tests that verify performance. For example, tests comparing the optimizers should show LASSO producing a larger number of terms than SR3 or STLSQ produces.
