### 2020-03-08 12855c1
Add property test for optimizers complexity
