### 2020-03-12 bfd9d41
fix notebook rendering with sphinx
