### 2020-03-14 #2104 @briandesilva
Thanks, Arfon, for the update. At the moment we **do not** need to put a pause on our involvement in the review process.
