### 2020-03-14 #2104 @sixpearls
Hi @briandesilva, I will be happy to do a follow up review when you are ready. I will need guidance on the authorship question.

@arfon could you advise on my authorship question?

> I am not sure I see the SOFTWARE contribution of the last three authors, although it seems that based on the references these authors may have contributed to the theoretical development. In the past, I believe the JOSS policy was relatively strict about the software contribution of each author. Is that still the case?
