### 2020-03-14 #2104 @arfon
@sixpearls - we generally try and steer clear of making authorship decisions. That said, we definitely do *not* require that all authors have to have contributed code/commits to the software as it's clearly possible for people to be important members of a project/collaboration without actually authoring the git commits.
