### 2020-03-31 df8068a
Add prose references to README
Add extra references for the original SINDy and SINDy-SR3 papers to the README.
