### 2020-04-02 82eacc7
Fix typo in stlsq.py
Fixed typo in ConvergenceWarning
