### 2020-04-03 #2104 @briandesilva
> * [ ]  It would greatly improve the experience for the reader to include at least some mathematical discussion and example code in the readme and setup.py `long_description` so that the benefits of the previous two suggests can be obtained for users without leaving github or PyPI.

@sixpearls, do you have any suggestions about how to render math in the README.rst file? I have drafted a simple mathematical description and some accompanying text for the README, but I have so far been unable to get any of the math to be rendered on GitHub.

After some poking around other repositories, the only solution I have been able to find is to export the rendered equations as images, then embed the images in the README (something like what [this repository](https://github.com/artivis/manif) does. This would work for users viewing the README on GitHub, but I'm not sure whether it would work for those viewing the package on PyPI.

Any recommendations you can provide on how to proceed would be greatly appreciated.
