### 2020-04-05 #2104 @sixpearls
@briandesilva I can't find a better example, but I was imagining just a simple [ASCII formulation like this one](https://github.com/cvxgrp/dccp#dccp-rules). You could alternatively describe the mathematical concepts "in" Python or pseudocode.

That said, Your update to the ["original paper" notebook](https://pysindy.readthedocs.io/en/latest/examples/2_original_paper.html#main-ideas) is exactly what I had in mind. Since your API is fairly semantic, I think a brief prose description with a well commented example and a link to the "main idea" section of this notebook (or breaking it out onto its own page) would be sufficient for the readme and PyPi long description.
