### 2020-04-05 #2104 @sixpearls
I believe the PyPI long description can render images, but you would need to find a host to serve the graphic to PyPI and I'm not sure if GitHub does that. So, rather than using plots in the PyPI long description, you could do something like show the output of numpy's `isclose` function to verify the fit.
