+++ README.rst
@@ -5,6 +5,84 @@ PySINDy
 
 **PySINDy** is a sparse regression package with several implementations for the Sparse Identification of Nonlinear Dynamical systems (SINDy) method introduced in Brunton et al. (2016), including the unified optimization approach of Champion et al. (2019).
 
+System identification
+---------------------
+System identification refers to the process of leveraging measurement data to infer governing equations, in the form of dynamical systems, describing the data. Once discovered, these equations can make predictions about future states, can inform control inputs, or can enable the theoretical study using analytical techniques.
+Dynamical systems are a flexible, well-studied class of mathematical objects for modeling systems evolving in time.
+SINDy is a model discovery method which uses *sparse regression* to infer nonlinear dynamical systems from measurement data.
+The resulting models are inherently *interpretable* and *generalizable*.
+
+How it works
+^^^^^^^^^^^^
+Suppose, for some physical system of interest, we have measurements of state variables `x(t)` (a vector of length n) at different points in time. Examples of state variables include the position, velocity, or acceleration of objects; lift, drag, or angle of attack of aerodynamic objects; and concentrations of different chemical species. If we suspect that the system could be well-modeled by a dynamical system of the form::
+
+    x'(t) = f(x(t)),
+
+then we can use SINDy to learn `f(x)` from the data (`x'(t)` denotes the time derivative of `x(t)`). Note that both `f(x)` and `x(t)` are typically vectors. The fundamental assumption SINDy employs is that each component of `f(x)`, `f_i(x)` can be represented as a *sparse* linear combination of basis functions `theta_j(x)`::
+
+  f_i(x) = theta_1(x) * xi_{1,i} + theta_2(x) * xi_{2,i} + ... + theta_l * xi{l,i}
+
+Concatenating all the objects into matrices (denoted with capitalized names) helps to simplify things.
+To this end we place all measurements of the state variables into a data matrix `X` (with a row per time measurement and a column per variable), the derivatives of the state variables into a matrix `X'`, all basis functions evaluated at all points in time into a matrix `Theta(X)` (each basis function gets a column), and all coefficients into a third matrix `Xi` (one column per state variable).
+The approximation problem to be solved can then be compactly written as::
+
+    X' = Theta(X) * Xi.
+
+Each row of this matrix equation corresponds to one coordinate function of `f(x)`.
+SINDy employs sparse regression techniques to find a solution `Xi` with sparse column vectors.
+For a more in-depth look at the mathematical foundations of SINDy, please see our introduction to SINDy `documentation <https://pysindy.readthedocs.io/en/latest/examples/2_introduction_to_sindy.html>`_.
+
+Relation to PySINDy
+^^^^^^^^^^^^^^^^^^^
+The PySINDy class revolves around the `SINDy` class which consists of three primary components; one for each term in the above matrix approximation problem.
+
+* `differentiation_method`: computes `X'`, though if derivatives are known or measured directly, they can be used instead
+* `feature_library`: specifies the candidate basis functions to be used to construct `Theta(X)`
+* `optimizer`: implements a sparse regression method for solving for `Xi`
+
+Once a `SINDy` object has been created it must be fit to measurement data, similar to a `scikit-learn` model. It can then be used to predict derivatives given new measurements, evolve novel initial conditions forward in time, and more. PySINDy has been written to be as compatible with `scikit-learn` objects and methods as possible.
+
+Example
[…]
