### 2020-04-10 #2104 @briandesilva
@sixpearls, I've added some information on the mathematics behind SINDy and a short example to the README (and the PyPI description). I also took your advice and moved the introduction to SINDY to its own notebook. Please take a look when you get a chance.
