+++ README.rst
@@ -30,7 +30,7 @@ The approximation problem to be solved can then be compactly written as::
 
 Each row of this matrix equation corresponds to one coordinate function of ``f(x)``.
 SINDy employs sparse regression techniques to find a solution ``Xi`` with sparse column vectors.
-For a more in-depth look at the mathematical foundations of SINDy, please see our introduction to SINDy `documentation <https://pysindy.readthedocs.io/en/latest/examples/2_introduction_to_sindy.html>`_.
+For a more in-depth look at the mathematical foundations of SINDy, please see our `introduction to SINDy <https://pysindy.readthedocs.io/en/latest/examples/2_introduction_to_sindy.html>`_.
 
 Relation to PySINDy
 ^^^^^^^^^^^^^^^^^^^
