### 2020-04-21 e756dab
Add tips page to documentation
