### 2020-04-23 f00fb7a
Update README.rst
Change number of basis functions from "l" to "k" for improved readability.
