### 2020-04-24 d3d9511
Update README.rst
Swap jupyter notebook example link to readthedocs link
