### 2020-04-24 f0fa8ec
Fix comment typo in utils
