### 2020-04-26 86b0efb
Fixed SINDy __init__ defaults
