### 2020-04-26 0a2fd46
Warning messages formatting
