### 2020-04-26 5c45a36
enable CI for pull requests
