### 2020-04-26 6b99173
Apply suggestions from code review
Superficial rearranging of code to appease black formatter
