### 2020-04-26 f9ac15b
Update pysindy/optimizers/stlsq.py
Fix formatting for black
