+++ README.rst
@@ -126,17 +126,29 @@ Community guidelines
 
 Contributing code
 ^^^^^^^^^^^^^^^^^
-We welcome contributions to PySINDy. To contribute a new feature please submit a pull request. To be accepted your code should conform to PEP8 (you may choose to use flake8 to test this before submitting your pull request). Your contributed code should pass all unit tests. Upon submission of a pull request, your code will be linted and tested automatically, but you may also choose to lint it yourself invoking
+We welcome contributions to PySINDy. To contribute a new feature please submit a pull request. To get started we recommend installing the packages in `requirements-dev.txt` via
 
 .. code-block:: bash
 
-  pre-commit -a -v
+    pip install -r requirements-dev.txt
 
-as well as test it yourself by running
+This will allow you to run unit tests and automatically format your code. To be accepted your code should conform to PEP8 and pass all unit tests. Code can be tested by invoking
 
 .. code-block:: bash
 
-  pytest
+    pytest
+
+We recommed using `pre-commit` to format your code. Once you have staged changes to commit
+
+.. code-block:: bash
+
+    git add path/to/changed/file.py
+
+you can run the following to automatically reformat your staged code
+
+.. code-block:: bash
+
+    pre-commit -a -v
 
 Reporting issues or bugs
 ^^^^^^^^^^^^^^^^^^^^^^^^
