### 2020-04-27 df3407a
Create functions to validate control inputs
