### 2020-04-27 305bd30
[WIP] Initial SINDYc setup
