### 2020-04-27 8392b4d
Fix typo in discrete time implementation
