### 2020-04-28 4ea222e
Refactor control variable validation functions
