### 2020-04-28 4d8bb25
Implement SINDyC class
