+++ pysindy/pysindy.py
@@ -116,14 +116,7 @@ class SINDy(BaseEstimator):
         self.n_jobs = n_jobs
 
     def fit(
-        self,
-        x,
-        t=1,
-        x_dot=None,
-        multiple_trajectories=False,
-        unbias=True,
-        control_variables=None,
-        quiet=False,
+        self, x, t=1, x_dot=None, multiple_trajectories=False, unbias=True, quiet=False,
     ):
         """
         Fit the SINDy model.
@@ -171,13 +164,6 @@ class SINDy(BaseEstimator):
             identified by the optimizer. This helps to remove the bias introduced by
             regularization.
 
-        control_variables: array-like or list of array-like, shape \
-                (n_samples, n_control_features)
-            Control variables. If training data contains multiple trajectories
-            (i.e. if x is a list of array-like), then control_variables should be a
-            list containing control variable data for each trajectory. Individual
-            trajectories may contain different numbers of samples
-
         quiet: boolean, optional (default False)
             Whether or not to suppress warnings during model fitting.
 
@@ -185,12 +171,6 @@ class SINDy(BaseEstimator):
         -------
         self: returns an instance of self
         """
-
-        if control_variables:
-            control_variables, self.control_indices = validate_control_variables(
-                x, control_variables, multiple_trajectories
-            )
-
         if multiple_trajectories:
             x, x_dot = self.process_multiple_trajectories(x, t, x_dot)
         else:
@@ -207,10 +187,6 @@ class SINDy(BaseEstimator):
                     x_dot = self.differentiation_method(x, t)
                 else:
                     x_dot = validate_input(x_dot, t)
-
-        # Append control variables
-        if control_variables:
-            x = concatenate((x, control_variables), axis=1)
 
         # Drop rows where derivative isn't known
         x, x_dot = drop_nan_rows(x, x_dot)
@@ -532,3 +508,295 @@ class SINDy(BaseEstimator):
     @property
     def complexity(self):
         return self.model.steps[-1][1].complexity
+
+
+class SINDyC(SINDy):
+    """
+    SINDyC (SINDy with control) model object.
+
+    Parameters
+    ----------
+    optimizer : optimizer object, optional
+        Optimization method used to fit the SINDy model. This must be an object
+        that extends the sindy.optimizers.BaseOptimizer class. Default is
+        sequentially thresholded least squares with a threshold of 0.1.
+
+    feature_library : feature library object, optional
+        Default is polynomial features of degree 2.
+
+    differentiation_method : differentiation object, optional
+        Method for differentiating the data. This must be an object that
+        extends the sindy.differentiation_methods.BaseDifferentiation class.
+        Default is centered difference.
+
+    feature_names : list of string, length n_input_features, optional
+        Names for the input features. If None, will use ['x0','x1',...].
+
+    discrete_time : boolean, optional (default False)
+        If True, dynamical system is treated as a map. Rather than predicting
+        derivatives, the right hand side functions step the system forward by
+        one time step. If False, dynamical system is assumed to be a flow
+        (right hand side functions predict continuous time derivatives).
+
+    n_jobs : int, optional (default 1)
+        The number of parallel jobs to use when fitting, predicting with, and
+        scoring the model.
+
+    Attributes
+    ----------
+    model : sklearn.multioutput.MultiOutputRegressor object
+        The fitted SINDy model.
+    """
+
+    def __init__(self, **kwargs):
+        super(SINDyC, self).__init__(**kwargs)
+
+    def fit(
+        self,
+        x,
+        u,
+        t=1,
+        x_dot=None,
+        multiple_trajectories=False,
+        unbias=True,
+        quiet=False,
+    ):
+        """
+        Fit the SINDy model.
+
+        Parameters
[…]
