### 2020-04-28 a0874e1
Handle 1D control input
