### 2020-04-28 49eb9f6
Fix SINDyC predict and score functions
