+++ pysindy/__init__.py
@@ -7,7 +7,8 @@ except DistributionNotFound:
     pass
 
 
-from pysindy.pysindy import SINDy
+from .pysindy import SINDy
+from .pysindy import SINDyC
 from pysindy.differentiation import *
 from pysindy.optimizers import *
 from pysindy.feature_library import *
