### 2020-04-28 66804e1
Add simulate function for SINDyC
