+++ pysindy/feature_library/__init__.py
@@ -2,3 +2,4 @@ from .custom_library import CustomLibrary
 from .fourier_library import FourierLibrary
 from .identity_library import IdentityLibrary
 from .polynomial_library import PolynomialLibrary
+from .concat_library import ConcatLibrary
