### 2020-04-28 b01eeae
Simulate with discrete time model
