### 2020-04-28 938b55d
Fix typo in multiple trajectories validation with control input
