### 2020-04-29 0b01f98
Allow for x to be a list
