+++ pysindy/utils/base.py
@@ -84,10 +84,10 @@ def _check_control_shape(x, u, trim_last_point):
         )
     if np.ndim(u) == 0:
         u = u[np.newaxis]
-    if x.shape[0] != u.shape[0]:
+    if len(x) != u.shape[0]:
         raise ValueError(
             "control variables u must have same number of rows as x. "
-            "u has {} rows and x has {} rows".format(u.shape[0], x.shape[0])
+            "u has {} rows and x has {} rows".format(u.shape[0], len(x))
         )
 
     if np.ndim(u) == 1:
