### 2020-04-29 a59ca6b
Refactor process_multiple_trajectories
