### 2020-04-29 ddae90e
Add unit tests for SINDyC
