+++ test/test_sindyc.py
@@ -1,0 +1,448 @@
+"""Unit tests for SINDyC: SINDy with control.
+
+SINDyC inherits from SINDy, so we focus on testing methods with
+new functionality.
+"""
+import numpy as np
+import pytest
+from scipy.integrate import odeint
+from sklearn.exceptions import ConvergenceWarning
+from sklearn.exceptions import NotFittedError
+from sklearn.linear_model import ElasticNet
+from sklearn.linear_model import Lasso
+from sklearn.utils.validation import check_is_fitted
+
+from pysindy import SINDyC
+from pysindy.optimizers import SR3
+from pysindy.optimizers import STLSQ
+
+
+@pytest.fixture
+def data_lorenz_c_1d():
+    def u(t):
+        return np.sin(2 * t)
+
+    def lorenz(z, t):
+        return [
+            10 * (z[1] - z[0]) + u(t) ** 2,
+            z[0] * (28 - z[2]) - z[1],
+            z[0] * z[1] - 8 / 3 * z[2],
+        ]
+
+    t = np.linspace(0, 5, 500)
+    x0 = [8, 27, -7]
+    x = odeint(lorenz, x0, t)
+    u_eval = u(t)
+
+    return x, t, u_eval, u
+
+
+@pytest.fixture
+def data_lorenz_c_2d():
+    def u(t):
+        return np.column_stack([np.sin(2 * t), t ** 2])
+
+    def lorenz(z, t):
+        u_eval = u(t)
+        return [
+            10 * (z[1] - z[0]) + u_eval[0, 0] ** 2,
+            z[0] * (28 - z[2]) - z[1],
+            z[0] * z[1] - 8 / 3 * z[2] - u_eval[0, 1],
+        ]
+
+    t = np.linspace(0, 5, 500)
+    x0 = [8, 27, -7]
+    x = odeint(lorenz, x0, t)
+    u_eval = u(t)
+
+    return x, t, u_eval, u
+
+
+@pytest.fixture
+def data_discrete_time_c():
+    def logistic_map(x, mu, ui):
+        return mu * x * (1 - x) + ui
+
+    n_steps = 100
+    mu = 3.6
+
+    u = 0.01 * np.random.randn(n_steps)
+    x = np.zeros((n_steps))
+    x[0] = 0.5
+    for i in range(1, n_steps):
+        x[i] = logistic_map(x[i - 1], mu, u[i - 1])
+
+    return x, u
+
+
+@pytest.fixture
+def data_discrete_time_multiple_trajectories_c():
+    def logistic_map(x, mu, ui):
+        return mu * x * (1 - x) + ui
+
+    n_steps = 100
+    mus = [1, 2.3, 3.6]
+    u = [0.001 * np.random.randn(n_steps) for mu in mus]
+    x = [np.zeros((n_steps)) for mu in mus]
+    for i, mu in enumerate(mus):
+        x[i][0] = 0.5
+        for k in range(1, n_steps):
+            x[i][k] = logistic_map(x[i][k - 1], mu, u[i][k - 1])
+
+    return x, u
+
+
+def test_get_feature_names_len(data_lorenz_c_1d):
+    x, t, u_eval, _ = data_lorenz_c_1d
+    model = SINDyC()
+
+    model.fit(x, u_eval, t=t)
+
+    # Assumes default library is polynomial features of degree 2
+    assert len(model.get_feature_names()) == 15
+
+
+def test_not_fitted(data_lorenz_c_1d):
+    x, t, u_eval, u = data_lorenz_c_1d
+    model = SINDyC()
+
+    with pytest.raises(NotFittedError):
+        model.predict(x, u_eval)
+    with pytest.raises(NotFittedError):
+        model.simulate(x[0], u, t)
+
+
[…]
