+++ test/feature_library/test_feature_library.py
@@ -8,11 +8,11 @@ from scipy.sparse import csr_matrix
 from sklearn.exceptions import NotFittedError
 from sklearn.utils.validation import check_is_fitted
 
+from pysindy.feature_library import ConcatLibrary
 from pysindy.feature_library import CustomLibrary
 from pysindy.feature_library import FourierLibrary
 from pysindy.feature_library import IdentityLibrary
 from pysindy.feature_library import PolynomialLibrary
-from pysindy.feature_library import ConcatLibrary
 from pysindy.feature_library.feature_library import BaseFeatureLibrary
 
 
