### 2020-04-29 f6d417c
Increase unit test coverage
