### 2020-04-29 cc7a5ea
Add example to documentation
