### 2020-04-29 6d8bd05
Clean up control variable validation
