### 2020-04-30 789a4b5
Move ConcatLibrary to feature_library.py and add a __add__ method to baseclass.
