+++ pysindy/feature_library/__init__.py
@@ -1,4 +1,4 @@
-from .concat_library import ConcatLibrary
+from .feature_library import ConcatLibrary
 from .custom_library import CustomLibrary
 from .fourier_library import FourierLibrary
 from .identity_library import IdentityLibrary
