+++ pysindy/feature_library/__init__.py
@@ -1,5 +1,5 @@
-from .feature_library import ConcatLibrary
 from .custom_library import CustomLibrary
+from .feature_library import ConcatLibrary
 from .fourier_library import FourierLibrary
 from .identity_library import IdentityLibrary
 from .polynomial_library import PolynomialLibrary
