### 2020-04-30 5d9cc8d
Adapt all tests for plus operator.
