### 2020-04-30 #2104 @dawbarton
Apologies for the delay, life has been a bit manic (and I also missed the ping somehow). I have been through the package, documentation, and paper - it looks good overall.

A few comments:

When I run the example in the README, I don't get the specified output. It's close but not quite the same; see below.

```
In [3]: model.print()
x' = -0.008 1 + -2.005 x + 0.002 y + 0.012 x y
y' = 1.000 y
```

Syntax highlighting in the docs is a little off - seems to get confused by ' characters.

Docs are relatively clear but slightly odd in places. For example, in the pysindy.feature_library package, the first class mentions things like `n_input_features\_`, and `n_output_features\_` but I can't see how these relate to the class (they don't appear to be parameters to the constructor or fields in the class - they are not valid identifiers). More of these terms ending with `\_` appear in the pysindy.optimizers package documentation.

The extensive examples are very nice. If only all packages had so many!

Overall, this is a great package. Thanks for putting it together.
