### 2020-04-30 #2104 @briandesilva
Thanks for your feedback!
- I will have to look into the README example. This looks like a bug -- those small magnitude terms should be getting ruled out.
- @Ohjeah, do you know of any fixes for the syntax highlighting issue? It looks like it's only rendering incorrectly on the sphinx site. We could change the way we write the math (e.g. write something like dX/dt), but for the example where we show the output of `model.print()`, we would have to change the call to something like `model.print(lhs=['dx/dt', 'dy/dt'])'. This makes the example look a little more complex than I would like.
- Let me provide some context for why we included the `n_input_features_` and `n_output_features_` attributes. When building our feature library objects we looked to scikit-learn's [PolynomialFeatures](https://scikit-learn.org/stable/modules/generated/sklearn.preprocessing.PolynomialFeatures.html) class, which has these attributes itself. They aren't known until the `fit` function is called because they depend on the shape of the input data, so they are automatically set then. Again, we are essentially just copying the structure of `PolynomialFeatures`. In any case, it looks like class attributes are not being rendered like I'd expect on the documentation site. They should be presented in a list under the heading "Attributes", but that's not how they're currently displayed. Sphinx is also adding extra backslashes before trailing underscores for these attributes. I will look into it.
