### 2020-04-30 dynamicslab/pysindy#73 @briandesilva
Attribute formatting in the documentation
The documentation site currently renders class attributes independently rather than grouping them in a list under the heading "Attributes". This issue is related to a comment in openjournals/joss-reviews/issues/2104.

Example (from `CustomLibrary`):
![Selection_001](https://user-images.githubusercontent.com/11828204/80754085-464ecc80-8ae3-11ea-99b1-6442b7289cba.png)

What we would like is for attributes to be displayed in a list, like the parameters in the above example.

It seems this is a [known issue](https://github.com/sphinx-doc/sphinx/issues/2115), and that one solution is to set `napoleon_use_ivar = True` in `docs/conf.py`, but this places the attributes under the heading "Variables" rather than "Attributes". Another [workaround](https://michaelgoerz.net/notes/extending-sphinx-napoleon-docstring-sections.html) appears to be to patch the napoleon package itself, but I'm not sure that would work for us because the documentation site is automatically built when changes are pushed.
