### 2020-04-30 bc13877
Lower default value of STLSQ.alpha
