+++ pysindy/optimizers/stlsq.py
@@ -23,7 +23,7 @@ class STLSQ(BaseOptimizer):
         Coefficients with magnitude below the threshold are set
         to zero.
 
-    alpha : float, optional (default 0.1)
+    alpha : float, optional (default 0.05)
         Optional L2 (ridge) regularization on the weight vector.
 
     max_iter : int, optional (default 20)
@@ -76,7 +76,7 @@ class STLSQ(BaseOptimizer):
     def __init__(
         self,
         threshold=0.1,
-        alpha=0.1,
+        alpha=0.05,
         max_iter=20,
         ridge_kw=None,
         normalize=False,
