### 2020-04-30 dynamicslab/pysindy#73 @briandesilva
Once we figure out how to do this properly, there are some missing attributes we should add to the `SINDy` class documentation, e.g. `n_input_features_` and `n_output_features_`.
