### 2020-05-01 cf72e18
Incorporate SINDyC into SINDy
