+++ pysindy/pysindy.py
@@ -91,6 +91,37 @@ class SINDy(BaseEstimator):
            [ 0.        ,  0.        ,  0.        ]])
     >>> model.score(x, t=t[1]-t[0])
     0.999999985520653
+
+    >>> import numpy as np
+    >>> from scipy.integrate import odeint
+    >>> from pysindy import SINDy
+    >>> u = lambda t : np.sin(2 * t)
+    >>> lorenz_c = lambda z,t : [
+                10 * (z[1] - z[0]) + u(t) ** 2,
+                z[0] * (28 - z[2]) - z[1],
+                z[0] * z[1] - 8 / 3 * z[2],
+        ]
+    >>> t = np.arange(0,2,0.002)
+    >>> x = odeint(lorenz_c, [-8,8,27], t)
+    >>> u_eval = u(t)
+    >>> model = SINDy()
+    >>> model.fit(x, u_eval, t=t[1]-t[0])
+    >>> model.print()
+    x0' = -10.000 x0 + 10.000 x1 + 1.001 u0^2
+    x1' = 27.994 x0 + -0.999 x1 + -1.000 x0 x2
+    x2' = -2.666 x2 + 1.000 x0 x1
+    >>> model.coefficients()
+    array([[ 0.        , -9.99969851,  9.99958359,  0.        ,  0.        ,
+             0.        ,  0.        ,  0.        ,  0.        ,  0.        ,
+             0.        ,  0.        ,  0.        ,  0.        ,  1.00120331],
+           [ 0.        , 27.9935177 , -0.99906375,  0.        ,  0.        ,
+             0.        ,  0.        , -0.99980455,  0.        ,  0.        ,
+             0.        ,  0.        ,  0.        ,  0.        ,  0.        ],
+           [ 0.        ,  0.        ,  0.        , -2.666437  ,  0.        ,
+             0.        ,  0.99990137,  0.        ,  0.        ,  0.        ,
+             0.        ,  0.        ,  0.        ,  0.        ,  0.        ]])
+    >>> model.score(x, u_eval, t=t[1]-t[0])
+    0.9999999855414495
     """
 
     def __init__(
@@ -116,7 +147,14 @@ class SINDy(BaseEstimator):
         self.n_jobs = n_jobs
 
     def fit(
-        self, x, t=1, x_dot=None, multiple_trajectories=False, unbias=True, quiet=False,
+        self,
+        x,
+        t=1,
+        x_dot=None,
+        u=None,
+        multiple_trajectories=False,
+        unbias=True,
+        quiet=False,
     ):
         """
         Fit the SINDy model.
@@ -147,6 +185,15 @@ class SINDy(BaseEstimator):
             provided, it must match the shape of the training data and these
             values will be used as the time derivatives.
 
+        u: array-like or list of array-like, shape (n_samples, n_control_features), \
+                optional (default None)
+            Control variables/inputs. Include this variable to use sparse
+            identification for nonlinear dynamical systems for control (SINDYc).
+            If training data contains multiple trajectories (i.e. if x is a list of
+            array-like), then u should be a list containing control variable data
+            for each trajectory. Individual trajectories may contain different
+            numbers of samples.
+
         multiple_trajectories: boolean, optional, (default False)
             Whether or not the training data includes multiple trajectories. If
             True, the training data must be a list of arrays containing data
@@ -171,6 +218,18 @@ class SINDy(BaseEstimator):
         -------
         self: returns an instance of self
         """
+        if u is None:
+            self.n_control_features_ = 0
+        else:
+            trim_last_point = self.discrete_time and (x_dot is None)
+            u = validate_control_variables(
+                x,
+                u,
+                multiple_trajectories=multiple_trajectories,
[…]
