### 2020-05-01 0e728d1
Update unit tests for SINDYc and add warning tests
