+++ test/test_sindyc.py
@@ -1,8 +1,4 @@
-"""Unit tests for SINDyC: SINDy with control.
-
-SINDyC inherits from SINDy, so we focus on testing methods with
-new functionality.
-"""
+"""Unit tests for SINDy with control."""
 import numpy as np
 import pytest
 from scipy.integrate import odeint
@@ -12,7 +8,7 @@ from sklearn.linear_model import ElasticNet
 from sklearn.linear_model import Lasso
 from sklearn.utils.validation import check_is_fitted
 
-from pysindy import SINDyC
+from pysindy import SINDy
 from pysindy.optimizers import SR3
 from pysindy.optimizers import STLSQ
 
@@ -94,9 +90,9 @@ def data_discrete_time_multiple_trajectories_c():
 
 def test_get_feature_names_len(data_lorenz_c_1d):
     x, t, u_eval, _ = data_lorenz_c_1d
-    model = SINDyC()
+    model = SINDy()
 
-    model.fit(x, u_eval, t=t)
+    model.fit(x, u=u_eval, t=t)
 
     # Assumes default library is polynomial features of degree 2
     assert len(model.get_feature_names()) == 15
@@ -104,12 +100,12 @@ def test_get_feature_names_len(data_lorenz_c_1d):
 
 def test_not_fitted(data_lorenz_c_1d):
     x, t, u_eval, u = data_lorenz_c_1d
-    model = SINDyC()
+    model = SINDy()
 
     with pytest.raises(NotFittedError):
-        model.predict(x, u_eval)
+        model.predict(x, u=u_eval)
     with pytest.raises(NotFittedError):
-        model.simulate(x[0], u, t)
+        model.simulate(x[0], t=t, u=u_eval)
 
 
 def test_improper_shape_input(data_1d):
@@ -117,33 +113,33 @@ def test_improper_shape_input(data_1d):
     u_eval = np.ones_like(x)
 
     # Ensure model successfully handles different data shapes
-    model = SINDyC()
-    model.fit(x.flatten(), u_eval, t=t)
+    model = SINDy()
+    model.fit(x.flatten(), u=u_eval, t=t)
     check_is_fitted(model)
 
-    model = SINDyC()
-    model.fit(x.flatten(), u_eval, t=t, x_dot=x.flatten())
+    model = SINDy()
+    model.fit(x.flatten(), u=u_eval, t=t, x_dot=x.flatten())
     check_is_fitted(model)
 
-    model = SINDyC()
-    model.fit(x, u_eval, t=t, x_dot=x.flatten())
+    model = SINDy()
+    model.fit(x, u=u_eval, t=t, x_dot=x.flatten())
     check_is_fitted(model)
 
-    model = SINDyC()
-    model.fit(x.flatten(), u_eval.reshape(-1, 1), t=t)
+    model = SINDy()
+    model.fit(x.flatten(), u=u_eval.reshape(-1, 1), t=t)
     check_is_fitted(model)
 
-    model = SINDyC()
-    model.fit(x.flatten(), u_eval.reshape(-1, 1), t=t, x_dot=x.flatten())
+    model = SINDy()
+    model.fit(x.flatten(), u=u_eval.reshape(-1, 1), t=t, x_dot=x.flatten())
     check_is_fitted(model)
 
-    model = SINDyC()
-    model.fit(x, u_eval.reshape(-1, 1), t=t, x_dot=x.flatten())
+    model = SINDy()
+    model.fit(x, u=u_eval.reshape(-1, 1), t=t, x_dot=x.flatten())
     check_is_fitted(model)
 
     # Should fail if x and u have incompatible numbers of rows
     with pytest.raises(ValueError):
[…]
