### 2020-05-01 c4834b2
Rename variables in SINDYc tests
