+++ test/test_sindyc.py
@@ -15,12 +15,12 @@ from pysindy.optimizers import STLSQ
 
 @pytest.fixture
 def data_lorenz_c_1d():
-    def u(t):
+    def u_fun(t):
         return np.sin(2 * t)
 
     def lorenz(z, t):
         return [
-            10 * (z[1] - z[0]) + u(t) ** 2,
+            10 * (z[1] - z[0]) + u_fun(t) ** 2,
             z[0] * (28 - z[2]) - z[1],
             z[0] * z[1] - 8 / 3 * z[2],
         ]
@@ -28,30 +28,30 @@ def data_lorenz_c_1d():
     t = np.linspace(0, 5, 500)
     x0 = [8, 27, -7]
     x = odeint(lorenz, x0, t)
-    u_eval = u(t)
+    u = u_fun(t)
 
-    return x, t, u_eval, u
+    return x, t, u, u_fun
 
 
 @pytest.fixture
 def data_lorenz_c_2d():
-    def u(t):
+    def u_fun(t):
         return np.column_stack([np.sin(2 * t), t ** 2])
 
     def lorenz(z, t):
-        u_eval = u(t)
+        u = u_fun(t)
         return [
-            10 * (z[1] - z[0]) + u_eval[0, 0] ** 2,
+            10 * (z[1] - z[0]) + u[0, 0] ** 2,
             z[0] * (28 - z[2]) - z[1],
-            z[0] * z[1] - 8 / 3 * z[2] - u_eval[0, 1],
+            z[0] * z[1] - 8 / 3 * z[2] - u[0, 1],
         ]
 
     t = np.linspace(0, 5, 500)
     x0 = [8, 27, -7]
     x = odeint(lorenz, x0, t)
-    u_eval = u(t)
+    u = u_fun(t)
 
-    return x, t, u_eval, u
+    return x, t, u, u_fun
 
 
 @pytest.fixture
@@ -89,57 +89,57 @@ def data_discrete_time_multiple_trajectories_c():
 
 
 def test_get_feature_names_len(data_lorenz_c_1d):
-    x, t, u_eval, _ = data_lorenz_c_1d
+    x, t, u, _ = data_lorenz_c_1d
     model = SINDy()
 
-    model.fit(x, u=u_eval, t=t)
+    model.fit(x, u=u, t=t)
 
     # Assumes default library is polynomial features of degree 2
     assert len(model.get_feature_names()) == 15
 
 
 def test_not_fitted(data_lorenz_c_1d):
-    x, t, u_eval, u = data_lorenz_c_1d
+    x, t, u, u_fun = data_lorenz_c_1d
     model = SINDy()
 
     with pytest.raises(NotFittedError):
-        model.predict(x, u=u_eval)
+        model.predict(x, u=u)
     with pytest.raises(NotFittedError):
-        model.simulate(x[0], t=t, u=u_eval)
+        model.simulate(x[0], t=t, u=u_fun)
 
 
 def test_improper_shape_input(data_1d):
     x, t = data_1d
-    u_eval = np.ones_like(x)
+    u = np.ones_like(x)
 
     # Ensure model successfully handles different data shapes
     model = SINDy()
-    model.fit(x.flatten(), u=u_eval, t=t)
+    model.fit(x.flatten(), u=u, t=t)
     check_is_fitted(model)
 
     model = SINDy()
-    model.fit(x.flatten(), u=u_eval, t=t, x_dot=x.flatten())
+    model.fit(x.flatten(), u=u, t=t, x_dot=x.flatten())
     check_is_fitted(model)
 
     model = SINDy()
-    model.fit(x, u=u_eval, t=t, x_dot=x.flatten())
[…]
