### 2020-05-01 #2104 @Ohjeah
Equation formatting in the readme is fixed now. The code-blocks were automatically interpreted as python and hence the funny looking syntax highlighting.
