### 2020-05-01 2fc4106
Relax optimizer complexity test
