### 2020-05-02 f8e30ef
Add n_output_features_ to ConcatLibrary
