+++ test/test_pysindy.py
@@ -188,6 +188,7 @@ def test_simulate(data):
         PolynomialLibrary(degree=3),
         FourierLibrary(n_frequencies=3),
         pytest.lazy_fixture("data_custom_library"),
+        PolynomialLibrary() + FourierLibrary(),
     ],
 )
 def test_libraries(data_lorenz, library):
