### 2020-05-04 dynamicslab/pysindy#73 @Ohjeah
The workaround looks good to me. However there are still some malformed sphinx roles causing lots of warnings/errors during the build:

```
/Users/mq/repo/github/dynamicslab/pysindy/docs/examples/2_original_paper.rst:36: WARNING: Unknown interpreted text role "raw-latex".
/Users/mq/repo/github/dynamicslab/pysindy/docs/examples/2_original_paper.rst:48: WARNING: Unknown interpreted text role "raw-latex".
/Users/mq/repo/github/dynamicslab/pysindy/docs/examples/2_original_paper.rst:69: WARNING: Unknown interpreted text role "raw-latex".
/Users/mq/repo/github/dynamicslab/pysindy/docs/examples/2_original_paper.rst:74: WARNING: Unknown interpreted text role "raw-latex".
/Users/mq/repo/github/dynamicslab/pysindy/docs/examples/2_original_paper.rst:79: WARNING: Unknown interpreted text role "raw-latex".
/Users/mq/repo/github/dynamicslab/pysindy/docs/examples/2_original_paper.rst:85: WARNING: Unknown interpreted text role "raw-latex".
/Users/mq/repo/github/dynamicslab/pysindy/docs/examples/2_original_paper.rst:100: WARNING: Unknown interpreted text role "raw-latex".

looking for now-outdated files... none found
pickling environment... /Users/mq/repo/github/dynamicslab/pysindy/docs/examples/2_original_paper.rst:108: WARNING: Unknown interpreted text role "raw-latex".
/Users/mq/repo/github/dynamicslab/pysindy/docs/examples/2_original_paper.rst:127: WARNING: Unknown interpreted text role "raw-latex".
/Users/mq/repo/github/dynamicslab/pysindy/docs/examples/2_original_paper.rst:253: WARNING: Unknown interpreted text role "raw-latex".
/Users/mq/repo/github/dynamicslab/pysindy/docs/examples/2_original_paper.rst:334: WARNING: Unknown interpreted text role "raw-latex".
/Users/mq/repo/github/dynamicslab/pysindy/docs/examples/2_original_paper.rst:404: WARNING: Unknown interpreted text role "raw-latex".
/Users/mq/repo/github/dynamicslab/pysindy/docs/examples/2_original_paper.rst:488: WARNING: Unknown interpreted text role "raw-latex".
/Users/mq/repo/github/dynamicslab/pysindy/docs/examples/2_original_paper.rst:787: WARNING: Unknown interpreted text role "raw-latex".
/Users/mq/repo/github/dynamicslab/pysindy/docs/examples/2_original_paper.rst:802: WARNING: Unknown interpreted text role "raw-latex".
/Users/mq/repo/github/dynamicslab/pysindy/docs/examples/2_original_paper.rst:807: WARNING: Unknown interpreted text role "raw-latex".
/Users/mq/repo/github/dynamicslab/pysindy/docs/examples/2_original_paper.rst:811: WARNING: Unknown interpreted text role "raw-latex".
/Users/mq/repo/github/dynamicslab/pysindy/docs/examples/2_original_paper.rst:815: WARNING: Unknown interpreted text role "raw-latex".
/Users/mq/repo/github/dynamicslab/pysindy/docs/examples/2_original_paper.rst:968: WARNING: Unknown interpreted text role "raw-latex".
/Users/mq/repo/github/dynamicslab/pysindy/docs/examples/2_original_paper.rst:976: WARNING: Unknown interpreted text role "raw-latex".
/Users/mq/repo/github/dynamicslab/pysindy/docs/examples/2_original_paper.rst: WARNING: document isn't included in any toctree
/Users/mq/repo/github/dynamicslab/pysindy/docs/examples/feature_overview.rst: WARNING: document isn't included in any toctree
/Users/mq/repo/github/dynamicslab/pysindy/docs/examples/papers/index.rst: WARNING: document isn't included in any toctree
/Users/mq/repo/github/dynamicslab/pysindy/docs/examples/papers/sindy_paper.rst: WARNING: document isn't included in any toctree
/Users/mq/repo/github/dynamicslab/pysindy/docs/examples/sindy_paper.rst: WARNING: document isn't included in any toctree
done
checking consistency... done
preparing documents... done
writing output... [100%] tips
generating indices...  genindex/Users/mq/repo/github/dynamicslab/pysindy/pysindy/pysindy.py:docstring of pysindy.pysindy.SINDy.fit:34: WARNING: 'any' reference target not found: SINDy.optimizer
/Users/mq/repo/github/dynamicslab/pysindy/pysindy/pysindy.py:docstring of pysindy.pysindy.SINDy.fit:34: WARNING: 'any' reference target not found: unbias=True
/Users/mq/repo/github/dynamicslab/pysindy/pysindy/pysindy.py:docstring of pysindy.pysindy.SINDy.predict:5: WARNING: 'any' reference target not found: multiple_trajectories=True
/Users/mq/repo/github/dynamicslab/pysindy/pysindy/pysindy.py:docstring of pysindy.pysindy.SINDy.score:14: WARNING: 'any' reference target not found: multiple_trajectories=True
/Users/mq/repo/github/dynamicslab/pysindy/pysindy/pysindy.py:docstring of pysindy.pysindy.SINDy.simulate:9: WARNING: 'any' reference target not found: self.discrete_time == False
/Users/mq/repo/github/dynamicslab/pysindy/pysindy/pysindy.py:docstring of pysindy.pysindy.SINDy.simulate:9: WARNING: 'any' reference target not found: self.discrete_time == True
/Users/mq/repo/github/dynamicslab/pysindy/pysindy/pysindy.py:docstring of pysindy.pysindy.SINDy.simulate:9: WARNING: 'any' reference target not found: len(u) == t
/Users/mq/repo/github/dynamicslab/pysindy/pysindy/pysindy.py:docstring of pysindy.pysindy.SINDy.simulate:9: WARNING: 'any' reference target not found: u.shape[0] == 1
/Users/mq/repo/github/dynamicslab/pysindy/pysindy/feature_library/polynomial_library.py:docstring of pysindy.feature_library.polynomial_library.PolynomialLibrary:8: WARNING: 'any' reference target not found: x[i] ** k
/Users/mq/repo/github/dynamicslab/pysindy/pysindy/feature_library/polynomial_library.py:docstring of pysindy.feature_library.polynomial_library.PolynomialLibrary:11: WARNING: 'any' reference target not found: degree
/Users/mq/repo/github/dynamicslab/pysindy/pysindy/feature_library/polynomial_library.py:docstring of pysindy.feature_library.polynomial_library.PolynomialLibrary:11: WARNING: 'any' reference target not found: x[1] ** 2
/Users/mq/repo/github/dynamicslab/pysindy/pysindy/feature_library/polynomial_library.py:docstring of pysindy.feature_library.polynomial_library.PolynomialLibrary:11: WARNING: 'any' reference target not found: x[0] * x[2] ** 3
/Users/mq/repo/github/dynamicslab/pysindy/pysindy/optimizers/sindy_optimizer.py:docstring of pysindy.optimizers.sindy_optimizer.SINDyOptimizer:3: WARNING: 'any' reference target not found: _unbias
/Users/mq/repo/github/dynamicslab/pysindy/pysindy/optimizers/sindy_optimizer.py:docstring of pysindy.optimizers.sindy_optimizer.SINDyOptimizer:7: WARNING: 'any' reference target not found: optimizer
/Users/mq/repo/github/dynamicslab/pysindy/pysindy/optimizers/sindy_optimizer.py:docstring of pysindy.optimizers.sindy_optimizer.SINDyOptimizer:7: WARNING: 'any' reference target not found: SINDyOptimizer.fit_intercept
/Users/mq/repo/github/dynamicslab/pysindy/pysindy/optimizers/sindy_optimizer.py:docstring of pysindy.optimizers.sindy_optimizer.SINDyOptimizer:7: WARNING: 'any' reference target not found: SINDyOptimizer.normalize
/Users/mq/repo/github/dynamicslab/pysindy/pysindy/utils/base.py:docstring of pysindy.utils.base.validate_control_variables:1: WARNING: 'any' reference target not found: return_array
/Users/mq/repo/github/dynamicslab/pysindy/pysindy/utils/base.py:docstring of pysindy.utils.base.validate_control_variables:1: WARNING: 'any' reference target not found: multiple_trajectories
../README.rst:191: WARNING: 'any' reference target not found: https://doi.org/10.1016/j.ifacol.2016.10.249
/Users/mq/repo/github/dynamicslab/pysindy/docs/tips.rst:4: WARNING: 'any' reference target not found: PySINDy
/Users/mq/repo/github/dynamicslab/pysindy/docs/tips.rst:4: WARNING: 'any' reference target not found: PySINDy
/Users/mq/repo/github/dynamicslab/pysindy/docs/tips.rst:29: WARNING: 'any' reference target not found: x_dot
/Users/mq/repo/github/dynamicslab/pysindy/docs/tips.rst:29: WARNING: 'any' reference target not found: differentiation_method
/Users/mq/repo/github/dynamicslab/pysindy/docs/tips.rst:29: WARNING: 'any' reference target not found: __call__
/Users/mq/repo/github/dynamicslab/pysindy/docs/tips.rst:29: WARNING: 'any' reference target not found: x
/Users/mq/repo/github/dynamicslab/pysindy/docs/tips.rst:29: WARNING: 'any' reference target not found: t
/Users/mq/repo/github/dynamicslab/pysindy/docs/tips.rst:38: WARNING: 'any' reference target not found: PySINDy
/Users/mq/repo/github/dynamicslab/pysindy/docs/tips.rst:38: WARNING: 'any' reference target not found: X
/Users/mq/repo/github/dynamicslab/pysindy/docs/tips.rst:49: WARNING: 'any' reference target not found: PySINDy
/Users/mq/repo/github/dynamicslab/pysindy/docs/tips.rst:51: WARNING: 'any' reference target not found: PySINDy
/Users/mq/repo/github/dynamicslab/pysindy/docs/tips.rst:53: WARNING: 'any' reference target not found: thresholder
/Users/mq/repo/github/dynamicslab/pysindy/docs/tips.rst:55: WARNING: 'any' reference target not found: optimizer
/Users/mq/repo/github/dynamicslab/pysindy/docs/tips.rst:55: WARNING: 'any' reference target not found: RidgeRegression
/Users/mq/repo/github/dynamicslab/pysindy/docs/tips.rst:55: WARNING: 'any' reference target not found: Lasso
/Users/mq/repo/github/dynamicslab/pysindy/docs/tips.rst:55: WARNING: 'any' reference target not found: ElasticNet
/Users/mq/repo/github/dynamicslab/pysindy/docs/tips.rst:55: WARNING: 'any' reference target not found: Lasso
/Users/mq/repo/github/dynamicslab/pysindy/docs/tips.rst:67: WARNING: 'any' reference target not found: unbias
/Users/mq/repo/github/dynamicslab/pysindy/docs/tips.rst:67: WARNING: 'any' reference target not found: True
```
