### 2020-05-04 dynamicslab/pysindy#73 @briandesilva
The "raw-latex" errors/warnings may be outdated as they are all related to `2_original_paper`, which was renamed to `3_original_paper`. When I renamed it I also thought I fixed all of those errors (#62, #64). I'll look into correcting all of those "any" warnings.
