### 2020-05-04 dynamicslab/pysindy#73 @briandesilva
I just pushed some updates to the docstrings mentioned in the above warning messages to #77.
