### 2020-05-04 f6aff8b
Reformat code in util docstring
