### 2020-05-04 b30047a
Add n_input_features_ and n_output_features_ attribute
