### 2020-05-04 7f74d0a
Add attributes to optimizers
