### 2020-05-04 #2104 @briandesilva
@dawbarton, we have updated the package and documentation to address each of your comments. When you get a chance could you please take another look to make sure everything meets your standards?
