### 2020-05-04 556b94c
Polish SINDy docstring
