### 2020-05-04 fe286f5
Enable interp functions for simulating with control inputs
