### 2020-05-04 #2104 @dawbarton
I haven't checked whether you've fixed the regression with the readme example but everything else looks good to me. 👍
