### 2020-05-04 e010360
Remove extraneous print statement
