### 2020-05-05 dynamicslab/pysindy#73 @Ohjeah
You are right, I might have had an old pickled environment which caused some of those errors.
