### 2020-05-06 fbcc83d
Clarify contributing code section of README
