### 2020-05-07 e1e62c3
Starting my own branch containing the implementation of the SR3 algorithm with linear equality constraints, as well as my additions to the SINDy algorithms using variable thresholding and initial guesses to the optimization problem.
