+++ pysindy/optimizers/__init__.py
@@ -1,4 +1,5 @@
 from .base import BaseOptimizer
 from .sindy_optimizer import SINDyOptimizer
 from .sr3 import SR3
+from .sr3_enhanced import SR3Enhanced
 from .stlsq import STLSQ
