### 2020-05-07 c06163e
Adding in the constrained optimization algorithm files provided by Kathleen.
