+++ pysindy/optimizers/constrained_stlsq.py
@@ -1,0 +1,151 @@
+import warnings
+
+import numpy as np
+from sklearn.exceptions import ConvergenceWarning
+from sklearn.linear_model import ridge_regression
+
+from pysindy.optimizers import BaseOptimizer
+
+
+class constrained_STLSQ(BaseOptimizer):
+    """
+    Sequentially thresholded least squares algorithm.
+
+    Attempts to minimize the objective function
+    ||y - Xw||^2_2 + alpha * ||w||^2_2 by iteratively performing
+    least squares and masking out elements of the weight that are
+    below a given threshold.
+
+    subject to  Cw = d (equality constraints)
+
+    Parameters
+    ----------
+    threshold : float, optional (default 0.1)
+        Minimum magnitude for a coefficient in the weight vector.
+        Coefficients with magnitude below the threshold are set
+        to zero.
+
+    alpha : float, optional (default 0)
+        Optional L2 (ridge) regularization on the weight vector.
+
+    max_iter : int, optional (default 20)
+        Maximum iterations of the optimization algorithm.
+
+    ridge_kw : dict, optional
+        Optional keyword arguments to pass to the ridge regression.
+
+    Attributes
+    ----------
+    coef_ : array, shape (n_features,) or (n_targets, n_features)
+        Weight vector(s)
+
+    ind_ : array, shape (n_features,) or (n_targets, n_features)
+        Array of 0s and 1s indicating which coefficients of the
+        weight vector have not been masked out.
+
+    Examples
+    --------
+    >>> import numpy as np
+    >>> from scipy.integrate import odeint
+    >>> from pysindy import SINDy
+    >>> from pysindy.optimizers import constrained_STLSQ
+    >>> lorenz = lambda z,t : [10*(z[1] - z[0]),
+    >>>                        z[0]*(28 - z[2]) - z[1],
+    >>>                        z[0]*z[1] - 8/3*z[2]]
+    >>> t = np.arange(0,2,.002)
+    >>> x = odeint(lorenz, [-8,8,27], t)
+    >>> opt = constrained_STLSQ(threshold=.1, alpha=.5)
+    >>> model = SINDy(optimizer=opt)
+    >>> model.fit(x, t=t[1]-t[0])
+    >>> model.print()
+    x0' = -9.999 1 + 9.999 x0
+    x1' = 27.984 1 + -0.996 x0 + -1.000 1 x1
+    x2' = -2.666 x1 + 1.000 1 x0
+    """
+
+    def __init__(self, threshold=0.1, alpha=0.0, max_iter=20, ridge_kw=None, **kwargs):
+        super(constrained_STLSQ, self).__init__(**kwargs)
+
+        if threshold < 0:
+            raise ValueError("threshold cannot be negative")
+        if alpha < 0:
+            raise ValueError("alpha cannot be negative")
+        if max_iter <= 0:
+            raise ValueError("max_iter must be positive")
+
+        self.threshold = threshold
+        self.alpha = alpha
+        self.max_iter = max_iter
+        self.ridge_kw = ridge_kw
+
+    def _sparse_coefficients(self, dim, ind, coef, threshold):
+        """Perform thresholding of the weight vector(s)
+        """
+        c = np.zeros(dim)
+        c[ind] = coef
+        big_ind = np.abs(c) >= threshold
+        c[~big_ind] = 0
+        self.history_.append(c)
+        return c, big_ind
+
+    def _regress(self, x, y):
+        """Perform the ridge regression
+        """
+        kw = self.ridge_kw or {}
+        coef = ridge_regression(x, y, self.alpha, **kw)
+        self.iters += 1
+        return coef
+
+    def _no_change(self):
+        """Check if the coefficient mask has changed after thresholding
[…]
