### 2020-05-07 2a9ca41
manually adding in the SINDYc changes to pysindy/utils/base.py
