+++ pysindy/utils/base.py
@@ -36,6 +36,56 @@ def validate_input(x, t=T_DEFAULT):
 
     return x
 
+def validate_control_variables(
+    x, u, multiple_trajectories=False, trim_last_point=False, return_array=True
+):
+    """
+    Ensure that control variables u are compatible with the data x.
+    If ``return_array`` and ``multiple_trajectories`` are True, convert u from a list
+    into an array (of concatenated list entries).
+    """
+    if multiple_trajectories:
+        if not isinstance(x, Sequence):
+            raise ValueError("x must be a list when multiple_trajectories is True")
+        if not isinstance(u, Sequence):
+            raise ValueError("u must be a list when multiple_trajectories is True")
+        if len(x) != len(u):
+            raise ValueError(
+                "x and u must be lists of the same length when "
+                "multiple_trajectories is True"
+            )
+
+        u_arr = [_check_control_shape(xi, ui, trim_last_point) for xi, ui in zip(x, u)]
+
+        if return_array:
+            u_arr = np.vstack(u_arr)
+
+    else:
+        u_arr = _check_control_shape(x, u, trim_last_point)
+
+    return u_arr
+
+
+def _check_control_shape(x, u, trim_last_point):
+    """
+    Convert control variables u to np.array(dtype=float64) and compare
+    its shape against x. Assumes x is array-like.
+    """
+    try:
+        u = np.array(u, dtype="float64")
+    except TypeError as e:
+        raise e(
+            "control variables u could not be converted to np.ndarray(dtype=float64)"
+        )
+    if np.ndim(u) == 0:
+        u = u[np.newaxis]
+    if len(x) != u.shape[0]:
+        raise ValueError(
+            "control variables u must have same number of rows as x. "
+            "u has {} rows and x has {} rows".format(u.shape[0], len(x))
+        )
+
+    if np.ndim(u) == 1:
 
 def drop_nan_rows(x, x_dot):
     x = x[~np.isnan(x_dot).any(axis=1)]
