### 2020-05-07 69a5267
Okay, think I have this branch up to date with master, and running correctly. Next steps are to implement the matrix of thresholds and initial guesses in STLSQ and clean things up.
