### 2020-05-07 29067e4
Added in the initial guess matrix to all the constrained and unconstrained optimization algorithms
