+++ pysindy/optimizers/stlsq.py
@@ -52,6 +52,11 @@ class STLSQ(BaseOptimizer):
     copy_X : boolean, optional (default True)
         If True, X will be copied; else, it may be overwritten.
 
+    initial_guess : 2D numpy array of floats (default None)
+        If user does not pass this, the initial guess for the optimization is
+        a naive lstsq (see below). If passes, the optimization starts
+        with this matrix as the initial starting point. 
+
     Attributes
     ----------
     coef_ : array, shape (n_features,) or (n_targets, n_features)
@@ -94,6 +99,7 @@ class STLSQ(BaseOptimizer):
         normalize=False,
         fit_intercept=False,
         copy_X=True,
+        initial_guess=None,
     ):
         super(STLSQ, self).__init__(
             max_iter=max_iter,
@@ -110,6 +116,7 @@ class STLSQ(BaseOptimizer):
         self.threshold = threshold
         self.alpha = alpha
         self.ridge_kw = ridge_kw
+        self.initial_guess = initial_guess
 
     def _sparse_coefficients(self, dim, ind, coef, threshold):
         """Perform thresholding of the weight vector(s)"""
@@ -142,6 +149,9 @@ class STLSQ(BaseOptimizer):
         Assumes an initial guess for coefficients and support are saved in
         ``self.coef_`` and ``self.ind_``.
         """
+        if self.initial_guess is not None:
+            self.coef_ = self.initial_guess.T
+ 
         ind = self.ind_
         n_samples, n_features = x.shape
         n_targets = y.shape[1]
