+++ pysindy/optimizers/sr3.py
@@ -81,6 +81,11 @@ class SR3(BaseOptimizer):
     copy_X : boolean, optional (default True)
         If True, X will be copied; else, it may be overwritten.
 
+    initial_guess : 2D numpy array of floats (default optional)
+        If user does not pass this, the initial guess for the optimization is
+        a naive lstsq (see below). If passes, the optimization starts
+        with this matrix as the initial starting point. 
+
     Attributes
     ----------
     coef_ : array, shape (n_features,) or (n_targets, n_features)
@@ -129,6 +134,7 @@ class SR3(BaseOptimizer):
         normalize=False,
         fit_intercept=False,
         copy_X=True,
+        initial_guess=None,
     ):
         super(SR3, self).__init__(
             max_iter=max_iter,
@@ -225,6 +231,9 @@ class SR3(BaseOptimizer):
 
         Assumes initial guess for coefficients is stored in ``self.coef_``.
         """
+        if self.initial_guess is not None:
+            self.coef_ = self.initial_guess.T
+ 
         coef_sparse = self.coef_.T
         n_samples, n_features = x.shape
 
