+++ pysindy/optimizers/__init__.py
@@ -1,7 +1,7 @@
 from .base import BaseOptimizer
 from .sindy_optimizer import SINDyOptimizer
 from .sr3 import SR3
-from .sr3_enhanced import SR3Enhanced
+from .constrained_sr3 import constrained_SR3 
 from .stlsq import STLSQ
 
 __all__ = ["BaseOptimizer", "SINDyOptimizer", "SR3", "STLSQ"]
