### 2020-05-07 b77b31f
Added in the initial guess matrix to all the constrained and unconstrained optimization algorithms
