+++ pysindy/optimizers/__init__.py
@@ -1,5 +1,5 @@
 from .base import BaseOptimizer
 from .sindy_optimizer import SINDyOptimizer
 from .sr3 import SR3
-from .sr3_enhanced import SR3Enhanced
+from .constrained_sr3 import constrained_SR3 
 from .stlsq import STLSQ
