### 2020-05-07 94a95f0
Adding in a matrix of thresholds in a consistent way, where the objective function reflects the change, is proving harder than I thought. Will take a stab at this after fixing the fact that some of the unit tests arent working.
