### 2020-05-07 4d819d1
Unit tests now all passing. Once again the issue was that the base.py I grabbed from Kathleens work did not correctly merge in the changes from master.
