### 2020-05-07 60f1093
Added in the small datafiles needed to make a simple plasma physics example based on the POD Galerkin paper for MHD.
