+++ examples/data/plasmaphysics_example_singularValues.txt
@@ -1,0 +1,2490 @@
+301965.113114
+186059.732268
+9263.846711
+7241.103726
+2133.504337
+1903.732295
+781.320710
+737.817421
+597.588109
+393.066909
+345.058424
+306.794444
+287.253694
+256.635106
+218.030237
+191.789522
+173.157520
+166.062561
+131.908325
+116.957439
+86.634388
+80.368915
+77.205813
+73.410399
+65.578506
+61.644804
+58.049736
+47.281713
+42.937337
+39.849676
+37.186835
+35.593826
+32.991619
+32.009656
+31.583117
+30.675172
+29.918252
+25.709890
+22.834395
+21.027252
+19.750122
+19.482117
+18.562952
+18.455377
+17.896378
+17.403857
+16.875785
+16.186344
+15.638117
+14.806722
+13.150614
+13.078894
+12.337035
+10.722138
+10.508873
+9.911926
+9.872698
+9.511991
+9.296212
+9.033366
+8.372755
+8.332403
+7.699755
+7.450941
+6.920516
+6.857003
+6.390080
+6.248764
+6.175891
+6.076467
+5.733385
+5.432073
+5.298213
+4.964033
+4.716960
+4.619054
+4.438098
+4.356528
+4.206754
+4.085095
+4.045739
+3.757389
+3.727104
+3.690061
+3.552694
+3.512338
+3.463634
+3.351768
+3.207562
+3.193625
+3.013635
+2.977971
+2.939731
+2.805591
+2.755835
+2.728806
+2.603434
+2.511448
+2.429762
+2.288883
+2.230521
+2.144527
+2.119257
+2.083944
+2.041542
+1.955524
+1.908104
+1.867431
+1.847928
+1.764800
+1.727283
+1.703298
+1.687387
+1.641697
+1.615302
+1.603529
+1.581174
+1.555352
+1.526635
+1.490677
+1.431688
+1.412010
+1.323690
+1.317073
+1.267410
+1.242203
+1.222436
+1.182704
+1.172880
+1.161154
+1.134462
+1.127607
+1.099796
+1.067161
+1.044413
+1.029568
+1.022417
+0.988785
+0.982124
+0.947359
+0.941955
+0.897265
+0.895816
+0.869997
+0.867824
+0.840561
+0.833979
+0.811364
+0.795872
+0.788929
+0.778933
+0.760096
+0.748940
+0.736536
+0.719353
+0.706712
+0.687720
+0.678161
+0.669691
+0.656931
+0.643653
+0.632455
+0.622972
+0.605052
+0.599492
+0.586827
+0.581449
+0.576603
+0.567553
+0.556788
+0.548031
+0.536368
+0.529333
+0.519779
+0.514212
+0.507592
+0.504454
+0.490253
+0.484758
+0.479810
+0.468013
+0.460028
+0.455738
+0.451146
+0.445718
+0.440241
+0.425484
+0.420362
+0.417016
+0.404915
+0.400780
+0.395485
+0.387363
+0.380488
+0.378684
+0.375819
+0.371548
+0.364779
+0.358035
+0.354517
+0.351460
+0.347529
+0.343293
+0.336659
+0.330683
+0.326423
+0.317920
+0.315534
+0.310754
+0.308992
+0.307386
+0.305731
+0.297374
+0.292846
+0.290831
+0.283507
+0.279030
+0.277883
+0.269808
+0.266364
+0.264322
+0.261564
+0.258119
+0.254116
+0.251519
+0.248771
+0.243856
+0.240254
+0.238149
+0.236689
+0.234026
+0.231204
[…]
