### 2020-05-07 78169e8
Getting the thresholds working in constrained SR3 first, where I have the most experience.
