### 2020-05-07 f6008f9
getting pretty close to it working on constrained SR3. Double checking it on my run to see if it can reproduce what I was doing before.
