### 2020-05-08 6c3b0e2
Added a ipython notebook with my POD-Galerkin plasma example.
