### 2020-05-08 24c3618
Added a ipython notebook with my POD-Galerkin plasma example.
