### 2020-05-09 c6af8f5
Fix flake8 errors and change class names to conform to PEP8
