+++ pysindy/optimizers/__init__.py
@@ -1,5 +1,5 @@
 from .base import BaseOptimizer
+from .constrained_sr3 import ConstrainedSR3
 from .sindy_optimizer import SINDyOptimizer
 from .sr3 import SR3
-from .constrained_sr3 import constrained_SR3 
 from .stlsq import STLSQ
