### 2020-05-09 881fc5f
Fix flake8 errors and change class names to conform to PEP8
