+++ pysindy/optimizers/constrained_sr3.py
@@ -1,14 +1,17 @@
 import warnings
+
 import numpy as np
 from scipy.linalg import cho_factor
 from scipy.linalg import cho_solve
 from scipy.optimize import bisect
 from sklearn.exceptions import ConvergenceWarning
+
 from pysindy.optimizers import BaseOptimizer
 from pysindy.utils import get_prox
 from pysindy.utils import get_reg
 
-class constrained_SR3(BaseOptimizer):
+
+class ConstrainedSR3(BaseOptimizer):
     """
     Sparse relaxed regularized regression.
 
@@ -66,7 +69,7 @@ class constrained_SR3(BaseOptimizer):
     initial_guess : 2D numpy array of floats (default optional)
         If user does not pass this, the initial guess for the optimization is
         a naive lstsq (see below). If passes, the optimization starts
-        with this matrix as the initial starting point. 
+        with this matrix as the initial starting point.
 
     unbias : boolean, optional (default True)
         Whether to perform an extra step of unregularized linear regression to unbias
@@ -76,7 +79,7 @@ class constrained_SR3(BaseOptimizer):
         Setting `unbias=True` will trigger an additional step wherein the nonzero
         coefficients learned by the `STLSQ` object will be updated using an
         unregularized least-squares fit.
-    
+
     thresholds : 2D array of floats, optional (default None)
         Optional threshold matrix to provide different thresholds
         for different terms in the optimization process
@@ -129,7 +132,7 @@ class constrained_SR3(BaseOptimizer):
         initial_guess=None,
         thresholds=None,
     ):
-        super(constrained_SR3, self).__init__(
+        super(ConstrainedSR3, self).__init__(
             max_iter=max_iter,
             normalize=normalize,
             fit_intercept=fit_intercept,
@@ -152,11 +155,15 @@ class constrained_SR3(BaseOptimizer):
         self.prox = get_prox(thresholder)
         self.reg = get_reg(thresholder)
         if thresholder[0].lower() == "w" and thresholds is None:
-            raise ValueError("weighted thresholder can only be used with a matrix of thresholds")
+            raise ValueError(
+                "weighted thresholder can only be used with a matrix of thresholds"
+            )
         if thresholder[0].lower() != "w" and thresholds is not None:
             raise ValueError("matrix of thresholds requires a weighted thresholder")
         if thresholds is not None and not np.all(thresholds >= 0.0):
-            raise ValueError("matrix of thresholds must contain non-negative entries only")
+            raise ValueError(
+                "matrix of thresholds must contain non-negative entries only"
+            )
         if trimming_fraction == 0.0:
             self.use_trimming = False
         else:
@@ -164,7 +171,9 @@ class constrained_SR3(BaseOptimizer):
             self.trimming_fraction = trimming_fraction
             self.trimming_initialization = trimming_initialization
             self.trimming_step_size = trimming_step_size
-        self.use_constraints = (constraint_lhs is not None) and (constraint_rhs is not None)
+        self.use_constraints = (constraint_lhs is not None) and (
+            constraint_rhs is not None
+        )
         if self.use_constraints:
             self.n_constraints = constraint_lhs.shape[0]
             self.constraint_lhs = constraint_lhs
@@ -190,10 +199,13 @@ class constrained_SR3(BaseOptimizer):
         g = x_transpose_y + coef_sparse / self.nu
         inv1 = np.linalg.inv(H)
         inv1_mod = np.kron(inv1, np.eye(coef_sparse.shape[1]))
[…]
