+++ pysindy/optimizers/sr3.py
@@ -84,7 +84,7 @@ class SR3(BaseOptimizer):
     initial_guess : 2D numpy array of floats (default optional)
         If user does not pass this, the initial guess for the optimization is
         a naive lstsq (see below). If passes, the optimization starts
-        with this matrix as the initial starting point. 
+        with this matrix as the initial starting point.
 
     Attributes
     ----------
@@ -233,7 +233,7 @@ class SR3(BaseOptimizer):
         """
         if self.initial_guess is not None:
             self.coef_ = self.initial_guess.T
- 
+
         coef_sparse = self.coef_.T
         n_samples, n_features = x.shape
 
