### 2020-05-09 39f69c4
Fix lingering flake8 error
