### 2020-05-09 077cfcf
Fix lingering flake8 error
