### 2020-05-11 e37c092
Update pysindy/optimizers/constrained_sr3.py
Co-authored-by: Markus Quade <markus.quade@gmail.com>
