+++ pysindy/optimizers/constrained_sr3.py
@@ -19,7 +19,7 @@ class ConstrainedSR3(BaseOptimizer):
 
     .. math::
 
-        0.5\\|y-Xw\\|^2_2 + lambda \\times R(v)
+        0.5\\|y-Xw\\|^2_2 + \\lambda \\times R(v)
         + (0.5 / nu)\\|w-v\\|^2_2
 
     where :math:`R(v)` is a regularization function. See the following reference
