### 2020-05-12 68310ba
Update README wording
