### 2020-05-13 cce6242
Define constant for coef_ thresholding
