### 2020-05-14 727ede8
Fix documentation typos
