### 2020-05-14 cee8b88
Minor improvements to constrained optimizers
