+++ pysindy/optimizers/__init__.py
@@ -1,5 +1,6 @@
 from .base import BaseOptimizer
 from .constrained_sr3 import ConstrainedSR3
+from .constrained_stlsq import ConstrainedSTLSQ
 from .sindy_optimizer import SINDyOptimizer
 from .sr3 import SR3
 from .stlsq import STLSQ
