### 2020-05-14 50da326
Minor improvements to constrained optimizers
