### 2020-05-15 a2e6280
Added some commenting and reshaped a few things. Changed the plasma example to be consisten with the new shapes. In the next commit, I will delete constraint_stlsq since this is not functional yet. For now, constrained SINDy users can utilize the constrained SR3 algorithm, which has been tested and works.
