+++ pysindy/optimizers/constrained_stlsq.py
@@ -127,7 +127,7 @@ class ConstrainedSTLSQ(BaseOptimizer):
         self.coef_ and self.ind_
         """
         if self.initial_guess is not None:
-            self.coef_ = self.initial_guess.T
+            self.coef_ = self.initial_guess
 
         ind = self.ind_
         n_samples, n_features = x.shape
