### 2020-05-15 7f537a4
Deleted constraint_stlsq
