### 2020-05-18 #2104 @briandesilva
@arfon - I have made a new release including all the changes prompted by this review. Here is the Zenodo DOI for PySINDy: [![DOI](https://zenodo.org/badge/186055899.svg)](https://zenodo.org/badge/latestdoi/186055899)
